Changed long long -> int64_t (for library)

+++ cphelper/library/cpp/lib/numeric/matrix/matrix.cpp
@@ -73,7 +73,7 @@ public:
         return m;
     }
 
-    friend matrix pow(matrix m, long long p) {
+    friend matrix pow(matrix m, int64_t p) {
         assert(m.rows == m.cols && p >= 0);
         matrix res = make_id(m.rows);
         for (; p > 0; p /= 2, m *= m) if (p & 1) res *= m;
